1.00_β08
MotionSupporter_1.00_β08 miumiu
・多段分割
　・キー間の統合を試した後に、変曲点統合を試すよう処理変更
　・オフセットなし
　・プルダウンをコンボボックスに変更。入力値で対象ボーンをリストアップ
・多段統合
　・やや閾値を小さめにしてキー統合
　・ボーンダイアログを多段分割と合併
・モデル名表示追加
