+++ motion_supporter64.spec
@@ -26,7 +26,7 @@ exe = EXE(pyz,
           a.zipfiles,
           a.datas,
           [],
-          name='MotionSupporter_1.00_β07_64bit',
+          name='MotionSupporter_1.00_β08_64bit',
           debug=False,
           bootloader_ignore_signals=False,
           strip=False,
