+++ src/executor.py
@@ -15,7 +15,7 @@ from utils.MLogger import MLogger
 from utils import MFileUtils
 from utils.MException import SizingException
 
-VERSION_NAME = "1.00_β06"
+VERSION_NAME = "1.00_β08"
 
 # 指数表記なし、有効小数点桁数6、30を超えると省略あり、一行の文字数200
 np.set_printoptions(suppress=True, precision=6, threshold=30, linewidth=200)
